fix: 쿼리 최적화

+++ src/main/kotlin/com/wafflestudio/csereal/core/notice/dto/NoticeSearchResponse.kt
@@ -1,7 +1,7 @@
 package com.wafflestudio.csereal.core.notice.dto
 
 data class NoticeSearchResponse(
-    val total: Long,
+    val total: Int,
     val searchList: List<NoticeSearchDto>
 ) {
